+++ src/Parser/Php/PhpArgument.php
@@ -8,6 +8,7 @@ final class PhpArgument
         private readonly ?string $name,
         private readonly ?PhpStringLiteral $stringLiteral,
         private readonly ?PhpCallable $callable = null,
+        private readonly ?string $expression = null,
     ) {
     }
 
@@ -24,5 +25,10 @@ final class PhpArgument
     public function callable(): ?PhpCallable
     {
         return $this->callable;
+    }
+
+    public function expression(): ?string
+    {
+        return $this->expression;
     }
 }